fix: SoundVariant's seed is now a long

+++ src/main/java/dev/dfonline/flint/actiondump/sound/SoundVariant.java
@@ -3,6 +3,6 @@ package dev.dfonline.flint.actiondump.sound;
 public record SoundVariant(
         String id,
         String name,
-        int seed
+        long seed
 ) {
 }
